requirements.txt
dj-database-url==0.5.0
Django==2.1.2
django-heroku==0.3.1
djangorestframework==3.8.2
gunicorn==19.9.0
postgres==2.2.2
psycopg2==2.7.5
psycopg2-binary==2.7.5
python-memcached==1.59
pytz==2018.5
six==1.11.0
whitenoise==4.1
